Extract hotspots configuration setup into custom hook
REDMINE-20853
